+++ index.html
@@ -52,6 +52,6 @@ SOFTWARE.
 
         
         <script src="./lib/base64js.min.js"></script>
-        <script src="./dist/bundle.js"></script>
+        <script src="./dist/main.bundle.js"></script>
     </body>
 </html>
