+++ webpack.config.js
@@ -1,8 +1,12 @@
+const path = require("path");
+
 module.exports = {
   entry: "./src/index.tsx",
   output: {
-      filename: "bundle.js",
-      path: __dirname + "/dist"
+      filename: "[name].bundle.js",
+      chunkFilename: "[name].bundle.js",
+      path: path.resolve(__dirname, "dist/"),
+      publicPath: "/dist/",
   },
 
   // Enable sourcemaps for debugging webpack's output.
